contact logic and tests fixed

+++ app/Models/Blacklist.ts
@@ -8,11 +8,11 @@ const BlacklistSchema = new Schema({
 
 BlacklistSchema.static("addToBlacklist", async function(userId: string, bannedId: string) {
     return await this.findOneAndUpdate({ user: userId, banned: bannedId },
-        { $set: { user: userId, banned: bannedId, added_at: Date.now() }}, { upsert: true, new: true });
+        { $set: { user: userId, banned: bannedId }, $setOnInsert: { added_at: Date.now() }}, { upsert: true, new: true });
 });
 
-BlacklistSchema.static("removeFromBlacklist", async function(id: string) {
-    return await this.deleteMany({ _id: id });
+BlacklistSchema.static("removeFromBlacklist", async function(id: string, userId: string) {
+    return await this.deleteMany({ _id: id, user: userId });
 });
 
 export default (mongoose: Mongoose) => mongoose.model("Blacklist", BlacklistSchema, "users.blacklist");
